feat: reorganizing part of the code and layout
