+++ path.go
@@ -6,7 +6,7 @@
 package fox
 
 import (
-	"cmp"
+	"github.com/tigerwill90/fox/internal/netutil"
 	"strings"
 )
 
@@ -176,46 +176,11 @@ func SplitHostPath(url string) (host, path string) {
 
 	// no path
 	if hostEnd < 0 {
-		host, _ = splitHostPort(url)
+		host, _ = netutil.SplitHostPort(url)
 		return host, "/"
 	}
 
-	host, _ = splitHostPort(url[:hostEnd])
-	path = cmp.Or(url[hostEnd:], "/")
+	host, _ = netutil.SplitHostPort(url[:hostEnd])
+	path = url[hostEnd:]
 	return
-}
-
-// splitHostPort separates host and port. If the port is not valid, it returns
-// the entire input as host, and it doesn't check the validity of the host.
-// Unlike net.SplitHostPort, but per RFC 3986, it requires ports to be numeric.
-func splitHostPort(hostPort string) (host, port string) {
-	host = hostPort
-
-	colon := strings.LastIndexByte(host, ':')
-	if colon != -1 && validOptionalPort(host[colon:]) {
-		host, port = host[:colon], host[colon+1:]
-	}
-
-	if strings.HasPrefix(host, "[") && strings.HasSuffix(host, "]") {
-		host = host[1 : len(host)-1]
-	}
-
-	return
-}
-
-// validOptionalPort reports whether port is either an empty string
-// or matches /^:\d*$/
-func validOptionalPort(port string) bool {
-	if port == "" {
-		return true
-	}
-	if port[0] != ':' {
-		return false
-	}
-	for _, b := range port[1:] {
-		if b < '0' || b > '9' {
-			return false
-		}
-	}
-	return true
 }
